desenvolvimento_django/pyproject.toml
[project]
name = "desenvolvimento-django"
version = "0.1.0"
description = ""
authors = [
    {name = "Kaua",email = "kauawt2@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django",
    "ipython"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest-django = "*"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "mysite.settings"